added structure & spaces/tabs reset
