+++ src/main/Main.java
@@ -1,0 +1,36 @@
+package main;
+import java.util.Arrays;
+
+import model.Customer;
+import model.MovieRental;
+import service.RentalInfo;
+
+/**
+ * @class Main
+ * 
+ * MISSION: it creates an information slip about movie rentals.
+ */
+public class Main {
+
+	/*
+	 * MAIN
+	 */
+	public static void main(String[] args) {
+		new Main().execute();
+	}
+
+	/*
+	 * PRIVATE METHODS
+	 */
+	private void execute() {
+		String expected = "Rental Record for C. U. Stomer\n\tYou've Got Mail\t3.5\n\tMatrix\t2.0\nAmount owed is 5.5\nYou earned 2 frequent points\n";
+
+		String result = new RentalInfo().statement(new Customer("C. U. Stomer", Arrays.asList(new MovieRental("F001", 3), new MovieRental("F002", 1))));
+
+		if (!result.equals(expected)) {
+			throw new AssertionError("Expected: " + System.lineSeparator() + String.format(expected) + System.lineSeparator() + System.lineSeparator() + "Got: " + System.lineSeparator() + result);
+		}
+
+		System.out.println("Success");
+	}
+}
