+++ src/model/MovieRental.java
@@ -1,12 +1,25 @@
+package model;
+
+/**
+ * @class MovieRental
+ * 
+ * MISSION: model for renting a movie.
+ */
 public class MovieRental {
     private String movieId;
     private int days;
-
+    
+    /*
+	 * CONSTRUCTOR
+	 */
     public MovieRental(String movieId, int days) {
         this.movieId = movieId;
         this.days = days;
     }
 
+    /*
+	 * GETTERS / SETTERS
+	 */
     public String getMovieId() {
         return movieId;
     }
